Added new page for exercise select
