+++ app/src/main/java/cs321/customstrength/addProgram.java
@@ -1,5 +1,6 @@
 package cs321.customstrength;
 
+import android.content.Intent;
 import android.support.v4.view.MenuItemCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -14,7 +15,7 @@ import android.widget.TextView;
 
 import java.util.ArrayList;
 
-public class addProgram extends AppCompatActivity implements SearchView.OnQueryTextListener {
+public class addProgram extends AppCompatActivity {
     TextView weekValue;
     LinearLayout weeks;
     TextView dayValue;
@@ -108,44 +109,15 @@ public class addProgram extends AppCompatActivity implements SearchView.OnQueryT
 
             Button addExercise = new Button(this);
             addExercise.setText("Add Exercise");
+            addExercise.setOnClickListener(new View.OnClickListener() {
+                public void onClick(View view) {
+                    Intent searchIntent=new Intent(view.getContext(), selectExercise.class);
+                    startActivity(searchIntent);
+                }
+            });
             currentDay.addView(addExercise);
 
             days.addView(currentDay);
         }
-    }
-
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_search, menu);
-
-        MenuItem searchItem = menu.findItem(R.id.search);
-        SearchView searchView = (SearchView) MenuItemCompat.getActionView(searchItem);
-        searchView.setOnQueryTextListener(this);
-
-        return true;
-    }
-
-    public boolean onQueryTextSubmit(String query) {
-        ArrayList<String> results = LoadExerciseData.searchExercises(query, LoadExerciseData.PRELOADED_EXERCISES);
-        LinearLayout displayResults = new LinearLayout(this);
-        for (int i = 0; i < results.size(); i++) {
-            TextView result = new TextView(this);
-            result.setText(results.get(i));
-            displayResults.addView(result);
-        }
-        if (results.size() == 0) {
-            TextView noResults = new TextView(this);
-            noResults.setText(query);
-            displayResults.addView(noResults);
-        }
-        mainLayout.addView(displayResults);
-        TextView test = new TextView(this);
-        test.setText(getApplicationInfo().dataDir);
-        mainLayout.addView(test);
-        return true;
-    }
-
-    public boolean onQueryTextChange(String newText) {
-        // User changed the text
-        return false;
     }
 }
